[feat] 추가기능 테스트 추가
